Preserve invariant in private API
This change ensures that the fingerprintAssetUrl function always
returns a string, even when the file isn’t found. Before, when the file
wasn’t found, it would return assetUrl as provided, which meant that it
would sometimes be a string (unfingerprinted) and sometimes an
unfingerprinted URL object.

This change shouldn’t be observable, since fingerprintAssetUrl is a
private function only called internally to add entries to the urlCache.
But, it keeps the nice invariant that the entries in urlCache are
always strings.

Because this change isn’t observable, there’s no need for a new test.
But all the old tests still pass.